+++ src/DonorProfiles/ServiceProvider.php
@@ -7,6 +7,7 @@ use Give\Helpers\Hooks;
 use Give\DonorProfiles\Shortcode as Shortcode;
 use Give\DonorProfiles\Block as Block;
 use Give\DonorProfiles\App as App;
+use Give\DonorProfiles\RequestHandler as RequestHandler;
 
 use Give\DonorProfiles\Profile as Profile;
 
@@ -14,12 +15,14 @@ use Give\DonorProfiles\Routes\LoginRoute;
 use Give\DonorProfiles\Routes\LogoutRoute;
 use Give\DonorProfiles\Routes\VerifyEmailRoute;
 
-use Give\DonorProfiles\Tabs\ProfileTab\Tab as ProfileTab;
 use Give\DonorProfiles\Tabs\DonationHistoryTab\Tab as DonationHistoryTab;
 use Give\DonorProfiles\Tabs\EditProfileTab\Tab as EditProfileTab;
 
 use Give\DonorProfiles\Tabs\TabsRegister;
 
+/**
+ * @since 2.10.0
+ */
 class ServiceProvider implements ServiceProviderInterface {
 
 	/**
@@ -36,16 +39,13 @@ class ServiceProvider implements ServiceProviderInterface {
 	 * @inheritDoc
 	 */
 	public function boot() {
+
+		Hooks::addAction( 'give_embed_head', App::class, 'loadAssets' );
+
+		Hooks::addFilter( 'query_vars', RequestHandler::class, 'filterQueryVars' );
+		Hooks::addAction( 'parse_request', RequestHandler::class, 'parseRequest' );
+
 		Hooks::addAction( 'init', Shortcode::class, 'addShortcode' );
-
-		// Register Tabs
-		Hooks::addAction( 'init', DonationHistoryTab::class, 'registerTab' );
-		Hooks::addAction( 'init', EditProfileTab::class, 'registerTab' );
-
-		Hooks::addAction( 'wp_enqueue_scripts', TabsRegister::class, 'enqueueTabAssets' );
-		Hooks::addAction( 'rest_api_init', TabsRegister::class, 'registerTabRoutes' );
-
-		Hooks::addAction( 'wp_enqueue_scripts', Shortcode::class, 'loadFrontendAssets' );
 
 		Hooks::addAction( 'rest_api_init', LoginRoute::class, 'registerRoute' );
 		Hooks::addAction( 'rest_api_init', LogoutRoute::class, 'registerRoute' );
@@ -56,8 +56,15 @@ class ServiceProvider implements ServiceProviderInterface {
 
 		if ( function_exists( 'register_block_type' ) ) {
 			Hooks::addAction( 'init', Block::class, 'addBlock' );
-			Hooks::addAction( 'wp_enqueue_scripts', Block::class, 'loadFrontendAssets' );
 			Hooks::addAction( 'enqueue_block_editor_assets', Block::class, 'loadEditorAssets' );
 		}
+
+		// Register Tabs
+		Hooks::addAction( 'init', DonationHistoryTab::class, 'registerTab' );
+		Hooks::addAction( 'init', EditProfileTab::class, 'registerTab' );
+
+		Hooks::addAction( 'give_embed_head', TabsRegister::class, 'enqueueTabAssets' );
+		Hooks::addAction( 'rest_api_init', TabsRegister::class, 'registerTabRoutes' );
+
 	}
 }