+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Edit Profile tab now persists data (#5486)
 -   Recurring Donations tab UI now connects to addon (#5584)
 -   Donor Profile authentication logic is now implemented (#5569)
+-   Admin can now set Donor Profile accent color (#5615)
 
 ### Changed
 -   Donor Profile Donation History UI is now polished (#5566)